Remove suggestion suppressions and fixture
