--- .sonarcloud.properties
@@ -1,4 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2016-2026 Objectionary.com
-# SPDX-License-Identifier: MIT
-
-sonar.exclusions=eo-runtime/pom.xml
