+++ eo-runtime/pom.xml
@@ -233,7 +233,7 @@
                 <!-- Required to be skipped, since we in runtime origin library. -->
                 <lint>prohibited-package</lint>
                 <!--
-                @todo #4148:30min. Enable `unused-void-attr` lint.
+                Tracked by issue #4148: enable `unused-void-attr` lint.
                       We should enable this lint after the following issue with magic
                       voids will be resolved: https://github.com/objectionary/lints/issues/574.
                       Don't forget to enable `unused-void-attr` in the `test–compile`
@@ -241,7 +241,7 @@
                 -->
                 <lint>unused-void-attr</lint>
                 <!--
-                @todo #4148:30min. Enable `duplicate-names-in-diff-context` lint.
+                Tracked by issue #4148: enable `duplicate-names-in-diff-context` lint.
                       For now its too pedantic.
                       See this: https://github.com/objectionary/lints/issues/578.
                       Also, we should fix the bug with false positives with lambda object
@@ -249,21 +249,21 @@
                 -->
                 <lint>duplicate-names-in-diff-context</lint>
                 <!--
-                @todo #4096:30min. Enable `sparse-decoration` lint.
+                Tracked by issue #4096: enable `sparse-decoration` lint.
                       After we merged EO tests together with their source objects, and removed
                       `+tests` meta, we have complains about sparse-decoration. Let's adjust and
                       enable it.
                 -->
                 <lint>sparse-decoration</lint>
                 <!--
-                @todo #4935:30min. Enable `comment-not-capitalized` lint.
-                      The lint incorrectly flags comments starting with `@to-do` as not capitalized.
+                Tracked by issue #4935: enable `comment-not-capitalized` lint.
+                      The lint incorrectly flags comments with puzzle markers as not capitalized.
                       We should enable it after the lints repo fixes the issue to
-                      ignore @to-do markers. See: https://github.com/objectionary/lints/issues/805
+                      ignore puzzle markers. See: https://github.com/objectionary/lints/issues/805
                 -->
                 <lint>comment-not-capitalized</lint>
                 <!--
-                @todo #5078:30min. Remove `idempotent-attribute-is-not-first` skip.
+                Tracked by issue #5078: remove `idempotent-attribute-is-not-first` skip.
                       The lint relies on the `xi-cactus` positional marker that the legacy
                       ANTLR parser emitted; the new spec-driven parser does not produce
                       this marker, so the lint now fires on every formation. Drop this
@@ -271,14 +271,14 @@
                 -->
                 <lint>idempotent-attribute-is-not-first</lint>
                 <!--
-                @todo #5221:60min Enable `wrong-test-order` lint.
+                Tracked by issue #5221: enable `wrong-test-order` lint.
                       The lint fires as a false positive when anonymous `>>` objects from
                       the main body are floated past tests by vars-float-up transformation.
                       Enable once objectionary/lints#954 is resolved.
                 -->
                 <lint>wrong-test-order</lint>
                 <!--
-                @todo #5221:30min Enable `compound-name` lint.
+                Tracked by issue #5221: enable `compound-name` lint.
                       Several objects in eo-runtime use compound names that follow system
                       or platform naming conventions (e.g. neg-inf, groups-count, sin-family).
                       Either rename the objects or adjust the lint rule, then enable it.
@@ -325,22 +325,6 @@
     </plugins>
   </build>
   <profiles>
-    <profile>
-      <id>suggestions-fixture</id>
-      <activation>
-        <property>
-          <name>!skipSuggestionsFixture</name>
-        </property>
-      </activation>
-      <dependencies>
-        <dependency>
-          <groupId>org.eolang</groupId>
-          <artifactId>eo-suggestions-fixture</artifactId>
-          <version>${project.version}</version>
-          <scope>test</scope>
-        </dependency>
-      </dependencies>
-    </profile>
     <profile>
       <id>jacoco</id>
       <build>
