--- eo-suggestions-fixture/src/main/java/org/eolang/eodeps/EOprinter.java
@@ -1,13 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2016-2026 Objectionary.com
- * SPDX-License-Identifier: MIT
- */
-
-package org.eolang.eodeps;
-
-/**
- * Dependency object used by runtime suggestion discovery tests.
- * @since 1.0
- */
-public interface EOprinter {
-}
