--- eo-suggestions-fixture/src/main/java/org/eolang/eodeps/package-info.java
@@ -1,10 +1,0 @@
-/*
- * SPDX-FileCopyrightText: Copyright (c) 2016-2026 Objectionary.com
- * SPDX-License-Identifier: MIT
- */
-
-/**
- * Dependency fixture package.
- * @since 1.0
- */
-package org.eolang.eodeps;
