+++ pom.xml
@@ -18,7 +18,6 @@
   <modules>
     <module>eo-parser</module>
     <module>eo-maven-plugin</module>
-    <module>eo-suggestions-fixture</module>
     <module>eo-runtime</module>
     <module>eo-integration-tests</module>
   </modules>
